Changes by noelelmquist

+++ core/src/main/java/com/IONA/TowerDefense/model/Waves.java
@@ -9,7 +9,7 @@ public class Waves {
     public static class Enemy {
         public String type;
         //public String difficulty;
-        public double delay;
+        public float delay;
     }
 
     public static class Wave {
